requirements.txt
fastapi[standard]
uvicorn[standard]
pandas
requests
google-generativeai
python-dotenv
pyjwt
pwdlib[argon2]